full/requirements.txt
openai>=1.0.0
rdflib>=6.0.0
openpyxl>=3.0.0
numpy>=1.24.0
scipy>=1.10.0
paho-mqtt>=1.6.0
sounddevice>=0.4.0
openai-whisper>=20231117
